mckesc-API/requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
Django==2.2
django-cors-headers==3.0.1
djangorestframework==3.9.4
gunicorn==19.9.0
idna==2.8
ndg-httpsclient==0.5.1
psycopg2==2.7.4
psycopg2-binary==2.8.2
pusher==2.1.3
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
pyOpenSSL==19.0.0
python-dateutil==2.7.3
python-memcached==1.59
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.2
whitenoise==3.3.0

